add stripe integration section
