+++ packages/stripe-app/stripe-app.json
@@ -1,6 +1,6 @@
 {
   "id": "com.example.dub",
-  "version": "0.0.20",
+  "version": "0.0.21",
   "name": "Dub",
   "icon": "",
   "permissions": [
